requirements.txt
alabaster==0.7.12
annoy==1.17.0
argcomplete==1.12.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.12.10
asttokens==2.0.8
attrs==22.1.0
autoflake==1.5.3
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.8.0
bleach==5.0.1
botocore==1.27.74
brotlipy==0.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
contourpy==1.0.6
coverage==6.4.4
cryptography==37.0.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
docutils==0.16
dparse==0.6.0
dpcpp-cpp-rt==2022.2.0
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.2
fonttools==4.38.0
gradio
idna==3.4
imagesize==1.4.1
importlib-metadata==4.12.0
iniconfig==1.1.1
intel-cmplr-lib-rt==2022.2.0
intel-cmplr-lic-rt==2022.2.0
intel-opencl-rt==2022.2.0
intel-openmp==2022.2.0
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
mkl==2022.2.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mypy==0.971
mypy-extensions==0.4.3
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.12
numpy==1.22.3
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipx==0.16.4
platformdirs==2.5.2
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.13.0
pylint==2.15.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pyupgrade==2.38.0
PyYAML==5.4.1
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.0
regex==2022.9.13
requests==2.28.1
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
safety==2.1.1
Send2Trash==1.8.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.5.0
tbb==2021.7.0
terminado==0.15.0
tinycss2==1.1.1
tokenize-rt==4.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
types-docutils==0.19.1
typing_extensions==4.3.0
urllib3==1.26.12
userpath==1.7.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==4.0.3
wincertstore==0.2
wrapt==1.14.1
zipp==3.8.1
